+++ src/main/util.ts
@@ -216,7 +216,6 @@ const deleteOldestVideo = (storageDir: any) => {
     const metadataPath = getMetadataFileForVideo(videoPath);
     const metadataJSON = fs.readFileSync(metadataPath);
     const metadata = JSON.parse(metadataJSON);
-    console.log(videoPath, metadata)
     return Boolean(metadata.protected);
 }  
 
@@ -280,6 +279,11 @@ const deleteOldestVideo = (storageDir: any) => {
     fs.writeFileSync(metadataFile, newMetadataJsonString);
 }
 
+// When packaged, we need to fix some paths
+const fixPathWhenPackaged = (p) => {
+    return p.replace("app.asar", "app.asar.unpacked");
+}
+
 export {
     getVideoState,
     writeMetadataFile,
@@ -287,5 +291,6 @@ export {
     isConfigReady,
     deleteVideo,
     openSystemExplorer,
-    toggleVideoProtected
+    toggleVideoProtected,
+    fixPathWhenPackaged
 };